New package: openacs-bootstrap5-theme
This is the initial checking of the OpenACS Bootstrap 5 Theme,
which is a port of the OpenACS Bootstrap 3 Theme.

In its current, premature state, the basic functionality is there for
the openacs_bootstrap5 (openacs_bootstrap5_turquois untouched), but it
covers already the base functionality of xowiki (longer way to go for
e.g. xowf, or the various other packages). Also, the graphical
fine-tuning has to be done, but e.g. glyphicons are already replaced
by Bootstrap Icons, etc.