Fix for sitebug #150 (security relevant)
The login widget at the navbar contained in its query variable
the password of a user in plain text (e.g. when the user has
mistyped the password, but as well on other error conditions).
Now, the password is filterer in the return_url in such cases.

Many thanks to Khy H for reporting this bug.

For details, see
https://openacs.org//bugtracker/openacs.org/bug?bug_number=150